requirements.txt
numpy
pillow
torch
torchvision
scikit-learn
flask
flask-cors